fix(person): the review's fix wave on the LLM matching phase
F1 The birth-year split now honours the LLM pairs. A year-less row can enter a set through a
pair ALONE, matches no bucket by name, and was dropped on the smallest year -- the wrong
person, and pairs_within then found only one side, so the row did not even record the match
that moved it. _pair_adjacency is built once per company and _split_by_birth_year tries the
pair relation before the name relation, keeping the smallest-year fallback for rows that
reach neither.

F2 A circuit breaker in run_match: once a page has attempted 20 calls, a failure share above
50% raises after the page's rows are written. A provider outage no longer finishes green
with an error row per company and the whole set to re-pay next run; the asset's RetryPolicy
supplies the backoff.

F3 Sticky errors. match_state_sql returns the hash AND the error for every stored company;
rate_limited: and http_error: are re-sent on the same input, invalid_response:, too many
candidates and unexpected: are not -- they are skipped, counted as the new skipped_sticky,
and write no state row, so their matched_at stops re-selecting the company for the fold.

F4 Short ordinal ids. The prompt renders candidates as c0..cN through the new
prompt_payload; parse_match_response maps them back. serialize_candidates keeps the full ids
and the members, so no stored input_hash moves (the value is now pinned in the tests).
max_tokens scales: max(4_000, 120 x candidates, profile.max_tokens).

F5 A truncated or empty answer is a paid call: CallResult carries finish_reason and _resolve
stores invalid_response: truncated/empty WITH the usage and the raw text.

F6 000399: confidence Float32 -> Float64, so the inclusive threshold comparison is exact;
and the input_hash paragraph corrected -- the hash is not in the ORDER BY, so a re-scored
pair replaces its row and what supersedes is the fold's join on the state row's hash.

F7 The batch's query-size comment carries the measured figures: 640 KB for the
double-bound match_pairs_sql, 1.64x headroom, ~320 KB for every other statement.

F8 The roles cap keeps the 20 MOST RECENT pairs, presented ascending.

F9 docs/person-design.md records the birthday-driven re-send cost, the sticky rule, the
circuit breaker and the Float64 inclusive threshold.

P1 _age takes the smallest age, not the first in row order. P2 _resolve records an
unexpected exception as one company's error instead of failing the run.

+++ corpscout/clickhouse/migrations/000399_corpscout_se_company_person_match.up.sql
@@ -8,11 +8,14 @@ CREATE DATABASE IF NOT EXISTS corpscout;
 -- parser, not by the engine, and the sort key is that ordered pair, so a re-match of the same
 -- pair replaces its row instead of adding a second.
 --
--- WHY input_hash IS ON BOTH TABLES. A re-match writes the company its new pairs and its new
--- state row. The previous input hash keeps its pair rows -- nothing deletes them -- and the
--- fold reads only the pairs whose input_hash equals the company state row hash, so the old
--- input is superseded rather than removed. The state table is one row per company
--- (ReplacingMergeTree ORDER BY company_id), which is what makes that comparison a single hash.
+-- WHY input_hash IS ON BOTH TABLES. It is NOT in this table's sort key, so it does not
+-- version anything: a re-match of the SAME pair replaces that pair's row, keeping the newer
+-- input's confidence, reason and hash. What the column does is certify. The fold reads only
+-- the pairs whose input_hash equals the company's state-row hash, so a row left behind by an
+-- input the company no longer has -- a pair the new candidate list no longer contains, or one
+-- the model stopped scoring -- is simply never read again, without anything deleting it. The
+-- state table is one row per company (ReplacingMergeTree ORDER BY company_id), which is what
+-- makes that comparison a single hash.
 --
 -- WHY raw_response IS STORED. The same reason the basic-info observation cache stores it: a
 -- paid answer is evidence, and a parse that changes must be re-readable against the exact text
@@ -40,7 +43,11 @@ CREATE TABLE IF NOT EXISTS corpscout.se_company_person_match
     source_b LowCardinality(String),
     name_a String,
     name_b String,
-    confidence Float32,
+    -- Float64, not Float32: the fold admits a pair with `confidence >= MATCH_THRESHOLD` and
+    -- that comparison is inclusive. Float32 would store 0.8 as 0.800000011920929 and a
+    -- threshold of 0.7 as 0.69999998807907104, so a pair scored at exactly the threshold
+    -- would be kept or dropped by the storage format rather than by the constant.
+    confidence Float64,
     reason String,
     model LowCardinality(String),
     prompt_version LowCardinality(String),
